--- src/config/constants.js
@@ -1,4 +1,0 @@
-export const Constants = {
-  BackGroundImage:
-    'https://firebasestorage.googleapis.com/v0/b/visualizer-new-devs-test.appspot.com/o/base.jpeg?alt=media&token=358ccdea-3cf9-4751-ae48-4631e4700554'
-};
