+++ src/config/constants/index.js
@@ -1,0 +1,17 @@
+const Constants = {
+  // Add common constants here
+  BackGroundImage:
+    'https://firebasestorage.googleapis.com/v0/b/visualizer-new-devs-test.appspot.com/o/base.jpeg?alt=media&token=358ccdea-3cf9-4751-ae48-4631e4700554',
+
+  // Add firebase constants here
+  firebase: {
+    apiKey: 'AIzaSyC_JwpXS4uj9sRRDrbFAtalE1QulNTmKnw',
+    authDomain: 'visualizer-new-devs-test.firebaseapp.com',
+    projectId: 'visualizer-new-devs-test',
+    storageBucket: 'visualizer-new-devs-test.appspot.com',
+    messagingSenderId: '702664185241',
+    appId: '1:702664185241:web:580752c50d570d0c89ef08'
+  }
+};
+
+export default Constants;
