+++ src/config/firebase/index.js
@@ -1,0 +1,12 @@
+import Constants from '../constants';
+
+const config = {
+  apiKey: Constants.firebase.apiKey,
+  authDomain: Constants.firebase.authDomain,
+  projectId: Constants.firebase.projectId,
+  storageBucket: Constants.firebase.storageBucket,
+  messagingSenderId: Constants.firebase.messagingSenderId,
+  appId: Constants.firebase.appId
+};
+
+export default config;
