+++ src/config/index.js
@@ -1,0 +1,3 @@
+export { default as constants } from './constants';
+export { default as firebase } from './firebase';
+export * from './firestore';
